# backend/internal/httpapi/panelsettings/docs/swagger.yaml
basePath: /api
definitions:
...
  shared.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
      message:
        example: invalid input
        type: string
    type: object
...
info:
  contact: {}
  description: Exodus dashboard and node-management API.
  license:
    name: AGPL-3.0
  title: Exodus API
  version: 3.0.0
paths:
...
  /subscription-settings:
    get:
      consumes:
      - application/json
      description: Get or update global subscription page and protocol settings
      parameters:
      - description: Subscription settings fields to patch
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage subscription settings
      tags:
      - Subscription Settings Controller
    patch:
      consumes:
      - application/json
      description: Get or update global subscription page and protocol settings
      parameters:
      - description: Subscription settings fields to patch
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage subscription settings
      tags:
      - Subscription Settings Controller
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"